* commit 'e07cd1238c6956b6dd3037ca597cf611d9c059a9':
  MB-29879: Allow sub-second values for defragmenter_interval
  MB-29858: ActiveStream: only request getNumItemsForCursor for existing cursors
  MB-29675: Move ObjectRegistry stats to CoreLocalStats
  MB-29816: Make comments in persistence cursor clearer
  MB-29675: Introduce CoreLocalStats class

+++ engines/ep/src/defragmenter.cc
@@ -145,7 +145,7 @@ std::chrono::microseconds DefragmenterTask::maxExpectedDuration() {
     return getChunkDuration() * 10;
 }
 
-size_t DefragmenterTask::getSleepTime() const {
+double DefragmenterTask::getSleepTime() const {
     return engine->getConfiguration().getDefragmenterInterval();
 }
